% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_box.R
\name{add_box}
\alias{add_box}
\title{Add nodes}
\usage{
add_box(
  prev_box = NULL,
  txt,
  just = c("center", "left", "right"),
  text_width = NULL,
  ...
)
}
\arguments{
\item{prev_box}{Previous node object, the created new node will be vertically
aligned with this node. Left this as `NULL` if this is the first node. The first
node will be aligned in the top center.}

\item{txt}{Text in the node. If the `prev_box` is a horizontally aligned multiple
nodes, a vector of with the same length must be provided.}

\item{just}{The justification for the text: left, center or right.}

\item{text_width}{a positive integer giving the target column for wrapping
lines in the output. String will not be wrapped if not defined (default).
The \code{\link[stringi]{stri_wrap}} function will be used if \code{stringi}
package installed as it is a better options for non-Latin language, otherwise
 \code{\link[base]{strwrap}} will be used.}

\item{...}{Other parameters pass to \link{textbox},}
}
\value{
A \code{stard} object.
}
\description{
Create/add vertically aligned labeled nodes or side nodes.
}
\seealso{
\code{\link{add_side_box}} \code{\link{add_split}} \code{\link{textbox}} 
\code{\link{add_label_box}}
}
