% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label_box.R
\name{add_label_box}
\alias{add_label_box}
\title{Add a vertically aligned label nodes on the left side.}
\usage{
add_label_box(
  prev_box,
  txt,
  only_terminal = TRUE,
  col = col,
  fill = fill,
  fontfam = fontfam,
  just = c("center", "left", "right"),
  ...
)
}
\arguments{
\item{prev_box}{A completed diagram created with \code{add_box}, \code{add_side_box} etc.}

\item{txt}{Text in the node. If a character string is provided, the label will be aligned
 to the last box if a character is provided. If a named vector, the labels will align to
corresponding row of the node. And the names is the number indicating row number of box to
 horizontally align with and value is the text in the box.}

\item{only_terminal}{If the txt is only for the terminal box, default. Otherwise, the side box will
also be accounted for.}

\item{col}{Colour for labels (blue by default)}

\item{fill}{Fill for labels (white by default)}

\item{fontfam}{Font family for labels (Arial by default)}

\item{just}{The justification for the text: center (default), left or right.}

\item{...}{Other parameters pass to \link{textbox},}
}
\value{
A \code{stard} object.
}
\description{
In a stard diagram, this can be used to indicate different stage.
}
\seealso{
\code{\link{add_side_box}} \code{\link{add_split}} \code{\link{textbox}} 
\code{\link{add_box}}
}
