% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_split.R
\name{add_split}
\alias{add_split}
\title{Add a splitting box}
\usage{
add_split(prev_box, txt, text_width = NULL, ...)
}
\arguments{
\item{prev_box}{Previous node that the newly created split box will be aligned.}

\item{txt}{Text in the node. If the `prev_box` is a horizontally aligned multiple
nodes, a vector of with the same length must be provided.}

\item{text_width}{a positive integer giving the target column for wrapping
lines in the output. String will not be wrapped if not defined (default).
The \code{\link[stringi]{stri_wrap}} function will be used if \code{stringi}
package installed as it is a better options for non-Latin language, otherwise
 \code{\link[base]{strwrap}} will be used.}

\item{...}{Other parameters pass to \link{textbox},}
}
\value{
A \code{stard.list} object.
}
\description{
This function will create a horizontally aligned nodes. The horizontal coordinate
will be automatically calculated if the coordinates not provided.
}
\seealso{
\code{\link{add_box}}, \code{\link{add_side_box}} \code{\link{textbox}}
}
