% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldat.R
\docType{data}
\name{mldat}
\alias{mldat}
\title{WID-qEC test methylight data (mldat)}
\format{
## `mldat`
A data frame with 602 rows and 72 columns with participant, phenotypic and methylation data
}
\source{
<https://ascopubs.org/doi/10.1200/JCO.22.00266>
}
\usage{
mldat
}
\description{
A dataset of the WID-qEC test
}
\keyword{datasets}
