% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stard.R
\docType{package}
\name{stard-package}
\alias{stard-package}
\title{Create STARD diagram}
\description{
To make it easy to create STARD diagrams for the transparent reporting on diagnostic accuracy studies.
This is done by creating a standardized disposition data, and using this data as the source for the creation
the STARD diagram. Manual changes are also possible to supply node text labels.
}
