% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stard_plot.R
\name{stard_plot}
\alias{stard_plot}
\title{Self generating STARD diagram}
\usage{
stard_plot(
  data,
  orders,
  side_box,
  index_test,
  type = NULL,
  side_note = NULL,
  labels = NULL,
  cex = 0.8,
  text_width = NULL,
  col = "#4F81BD",
  fill = "white",
  fontfam = "Helvetica"
)
}
\arguments{
\item{data}{Data set with study and diagnostic information for each participant, including those that were not ultimately included}

\item{orders}{A named vector or a list, names as the variable in the dataset
and values as labels in the box. The order of the diagram will be based on this.}

\item{side_box}{Variable vector, appeared as side box in the diagram. The next
box will be the subset of the missing values of these variables.}

\item{index_test}{Name of the index_test variable. The
diagram will split into branches on this variables forward.}

\item{type}{Final assigned diagnosis, to be included in the bottom box of the STARD diagram}

\item{side_note}{Variable vector, appeared as side box in the diagram. The difference to the
box is that the following boxes will not be subset.}

\item{labels}{Named vector, names is the location of the terminal node. The
position location should plus 1 after the allocation variables if the allocation
is defined.}

\item{cex}{Multiplier applied to font size, Default is 0.8#'}

\item{text_width}{a positive integer giving the target column for wrapping
lines in the output. String will not be wrapped if not defined (default).
The \code{\link[stringi]{stri_wrap}} function will be used if \code{stringi}
package installed, otherwise \code{\link[base]{strwrap}} will be used.}

\item{col}{Colour for labels (blue by default)}

\item{fill}{Fill for labels (white by default)}

\item{fontfam}{Font family for labels (Arial by default)}
}
\value{
A \code{STARD} object.
}
\description{
Create STARD diagram from study data.
}
\details{
The calculation of numbers is as in an analogous to Kirchhoff's Laws of
electricity. The numbers in terminal nodes must sum to those in the ancestor
 nodes. All the drop outs will be populated as a side box. Which was different
 from the official STARD diagram template, which has dropout inside a
  vertical node.
}
\examples{
library(dplyr)

# Prepare test data


# Load data
mldat <- stard::mldat

# Filter Barcelona Validation Set
data <- mldat |> 
  dplyr::filter(set == 'Barcelona Validation')

# Excluded participants with insufficient DNA were not included
# in the final dataframe, so rows are added here
excl <- data.frame(excluded = rep("Participant not eligible for WID-qEC", 38))
insufficient <- data.frame(insufficient = rep("Insufficient DNA", 40))
data$excluded <- NA
data$insufficient <- NA
data <- plyr::rbind.fill(data,
                         excl,
                         insufficient)

# Assign a unique pseudo-ID
data$id = 1:nrow(data)

# Assign a threshold and whether a histology was available
data <- data |> 
  dplyr::mutate(
    qEC = ifelse(sum > 0.03, "positive", "negative"),
    histology_available = case_when(controldiag == 'Hospital control' &
                                      type == 'Control' ~
                                      "No histology available (hospital control)",
                                    TRUE ~ NA))

# Additional text for formatting
data <- data |> 
  dplyr::mutate(excltext = ifelse(is.na(excluded), "Not eligible", NA),
                insufficienttext = ifelse(is.na(insufficient), "Not sufficient", NA))

data <- data |> 
  dplyr::select(id, excluded, excltext, insufficient, qEC, histology_available, type)

head(data)


out <- stard_plot(data = data,
                  order = c(id ="Population",
                            excluded = "Eligible participants",
                            excltext  = "Eligible participants",
                            insufficient = "Included participants",
                            qEC = "WID-qEC",
                            histology_available = 'No histology',
                            id = 'number',
                            type = "Assigned type"),
                  side_box = c("excluded",
                               "insufficient"),
                  side_note = 'histology_available',
                  index_test = "qEC",
                  type = 'type',
                  labels = c("1" = "Potentially eligible population",
                             "2" = "Eligible participants",
                             "3" = 'Index Test',
                             "4" = "Index test outcome",
                             "5" = "Reference allocation"),
                  cex = 0.8)


plot(out)
}
\seealso{
\code{\link{add_side_box}},\code{\link{add_split}},
\code{\link{add_side_box}} \code{\link{textbox}}
}
